fix tests on travis
